# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.authRequest:
    properties:
      password:
        type: string
      phoneNumber:
        type: string
    type: object
  auth.authResponse:
    properties:
      token:
        type: string
    type: object
  auth.registerRequest:
    properties:
      password:
        type: string
      phoneNumber:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
  domain.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  domain.Product:
    properties:
      barcode:
        type: string
      category:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      stocks:
        items:
          $ref: '#/definitions/domain.Stock'
        type: array
      updatedAt:
        type: string
    type: object
  domain.Role:
    properties:
      id:
        type: integer
      name:
        type: string
      users:
        items:
          $ref: '#/definitions/domain.User'
        type: array
    type: object
  domain.Stock:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      movements:
        items:
          $ref: '#/definitions/domain.StockMovement'
        type: array
      product:
        $ref: '#/definitions/domain.Product'
      productID:
        type: integer
      quantity:
        type: integer
      updatedAt:
        type: string
      warehouse:
        $ref: '#/definitions/domain.Warehouse'
      warehouseID:
        type: integer
    type: object
  domain.StockMovement:
    properties:
      createdAt:
        type: string
      createdBy:
        $ref: '#/definitions/domain.User'
      createdByID:
        type: integer
      description:
        type: string
      id:
        type: integer
      quantity:
        type: integer
      stock:
        $ref: '#/definitions/domain.Stock'
      stockID:
        type: integer
      type:
        description: '"incoming" | "outgoing" | "transfer"'
        type: string
    type: object
  domain.User:
    properties:
      full_name:
        type: string
      id:
        type: integer
      password:
        type: string
      phone_number:
        type: string
      roles:
        items:
          $ref: '#/definitions/domain.Role'
        type: array
    type: object
  domain.Warehouse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/domain.User'
      owner_id:
        type: integer
      stocks:
        items:
          $ref: '#/definitions/domain.Stock'
        type: array
      updatedAt:
        type: string
    type: object
  dto.ProductRequest:
    properties:
      barcode:
        type: string
      category:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
  dto.UserResponse:
    properties:
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Description
  title: Voice-app API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login and receive JWT token
      parameters:
      - description: Login data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.authRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.authResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /products:
    get:
      description: Get all products
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get products
      tags:
      - products
    post:
      description: Add new product
      parameters:
      - description: Product data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ProductRequest'
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add product
      tags:
      - products
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with phone, password and roles
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - auth
  /users:
    get:
      description: Get all users or filter by phone number
      parameters:
      - description: Phone number
        in: query
        name: phone_number
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get users
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"